Add JS Interpreter scripts to the Startup App list
