+++ src/core/menu_items/ScriptsMenu.h
@@ -7,7 +7,7 @@ class ScriptsMenu : public MenuItemInterface {
 public:
     ScriptsMenu() : MenuItemInterface("JS Interpreter") {}
 
-    void optionsMenu(void);
+    void optionsMenu();
     void drawIcon(float scale);
     bool hasTheme() { return bruceConfig.theme.interpreter; }
     String themePath() { return bruceConfig.theme.paths.interpreter; }
