+++ src/modules/bjs_interpreter/interpreter.h
@@ -17,5 +17,8 @@ void run_bjs_script();
 bool run_bjs_script_headless(char *code);
 bool run_bjs_script_headless(FS fs, String filename);
 
+String getScriptsFolder(FS *&fs);
+std::vector<Option> getScriptsOptionsList(bool saveStartupScript = false);
+
 #endif
 #endif
